+++ domains/Discussions/Policies/QuestionPolicy.php
@@ -2,6 +2,7 @@
 
 namespace Domains\Discussions\Policies;
 
+use Domains\Accounts\Enums\AccountTypeEnum;
 use Domains\Accounts\Models\User;
 use Domains\Discussions\Models\Question;
 use Illuminate\Auth\Access\HandlesAuthorization;
@@ -10,8 +11,13 @@ class QuestionPolicy
 {
     use HandlesAuthorization;
 
-    public function update(User $user, Question $question)
+    public function update(User $user, Question $question): bool
     {
         return $question->author->is($user);
+    }
+
+    public function delete(User $user, Question $question): bool
+    {
+        return $question->author->is($user) || $user->hasRole(AccountTypeEnum::ADMIN);
     }
 }